% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtroc-package.R
\docType{package}
\name{mtroc-package}
\alias{mtroc}
\alias{mtroc-package}
\title{mtroc: Meta Analysis of Diagnostic Accuracy with Missing Data}
\description{
Using a two-level random-effects model for multiple imputations of missing data in meta analysis of diagnostic accuracy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lmaowisc/mtroc}
  \item \url{https://lmaowisc.github.io/mtroc/}
  \item Report bugs at \url{https://github.com/lmaowisc/mtroc/issues}
}

}
\author{
\strong{Maintainer}: Lu Mao \email{lmao@biostat.wisc.edu} (\href{https://orcid.org/0000-0002-8626-9822}{ORCID})

}
\keyword{internal}
